refactor Simple search
