+++ app/frontend/stylesheets/object/component/simple-search-view.scss
